refactor: 비즈니스 모델의 view 의존성 제거
