+++ src/main/java/racingcar/domain/CarRacingGameEngine.java
@@ -1,7 +1,6 @@
 package racingcar.domain;
 
 import racingcar.util.RandomUtils;
-import racingcar.view.OutputView;
 
 import java.util.List;
 
@@ -10,25 +9,12 @@ public class CarRacingGameEngine {
     private static final int MIN_RANDOM_VALUE = 0;
     private static final int MAX_RANDOM_VALUE = 9;
 
-    private final int tryTime;
-
-    public CarRacingGameEngine(List<Car> cars, int tryTime) {
+    public CarRacingGameEngine(List<Car> cars) {
         CarRepository.saveAll(cars);
-        this.tryTime = tryTime;
     }
 
     public void race() {
-        List<Car> cars = CarRepository.cars();
-
-        OutputView.printRaceResult();
-
-        for (int i = 0; i < tryTime; i++) {
-            updateAllCarsPosition(cars);
-            showAllCarsPosition(cars);
-            OutputView.printNewLine();
-        }
-
-        OutputView.printWinner(CarRepository.winnerNames());
+        updateAllCarsPosition(CarRepository.cars());
     }
 
     private void updateAllCarsPosition(List<Car> cars) {
@@ -44,11 +30,5 @@ public class CarRacingGameEngine {
         if (GoForwardOrStop.isGoForward(randomNumber)) {
             car.goForward();
         }
-    }
-
-    private void showAllCarsPosition(List<Car> cars) {
-        cars.forEach(car ->
-                OutputView.printRaceResultEachCar(
-                        car.getName(), car.getPosition()));
     }
 }
